WIP breaking up server/client distributed actor systems
