+++ Sources/XPCDistributedActorSystem/GenericInvocationEncoder.swift
@@ -15,8 +15,22 @@ public struct GenericInvocationEncoder: DistributedTargetInvocationEncoder, Send
     var returnType: String? = nil
     var errorType: String? = nil
 
+    private let client: XPCDistributedActorClient?
+    private let system: XPCDistributedActorSystem?
+
+    init(client: XPCDistributedActorClient) {
+        self.client = client
+        self.system = nil
+    }
+    
+    init(system: XPCDistributedActorSystem) {
+        self.client = nil
+        self.system = system
+    }
+
     public mutating func recordArgument(_ argument: RemoteCallArgument<some Codable>) throws {
-        try arguments.append(Self.encoder.encode(argument.value))
+        let data = try Self.encoder.encode(argument.value)
+        arguments.append(data)
     }
 
     public mutating func recordGenericSubstitution(_ type: (some Any).Type) throws {
