+++ Tests/XPCDistributedActorSystemTests/XPCDistributedActorSystemTests.swift
@@ -8,8 +8,8 @@ import Testing
 distributed actor Calculator {
     typealias ActorSystem = XPCDistributedActorSystem
 
-    init(actorSystem: ActorSystem) {
-        self.actorSystem = actorSystem
+    init(actorSystemServer: XPCDistributedActorServer) {
+        actorSystem = actorSystemServer
     }
 
     distributed func add(_ value1: Int, _ value2: Int) -> Int {
@@ -58,31 +58,36 @@ struct XPCDistributedActorSystemTests {
         try await confirmation("creates remote actor just once") { confirmCreatesActor in
             let xpcHostDAS = try await XPCDistributedActorServer(
                 listener: listenerXPC,
-                actorCreationHandler: { system in
+                actorCreationHandler: { actorSystemServer in
                     // Create a Calculator actor when one isn't found for the given ID
                     confirmCreatesActor()
-                    return Calculator(actorSystem: system)
+                    return Calculator(actorSystemServer: actorSystemServer)
                 },
             )
+            do {
+                // Create a client actor system that connects to the listener's endpoint
+                let clientDAS = try await XPCDistributedActorClient(
+                    connectionType: .endpoint(listenerXPC.endpoint),
+                    codeSigningRequirement: nil,
+                )
 
-            // Create a client actor system that connects to the listener's endpoint
-            let clientDAS = try await XPCDistributedActorClient(
-                connectionType: .endpoint(listenerXPC.endpoint),
-                codeSigningRequirement: nil,
-            )
+                // For now, let's create a remote reference manually since resolve might return nil for remote actors
+                // We'll use the distributed actor initializer that takes an ID and system
+                let calculator = try Calculator.resolve(id: .init(), using: clientDAS)
+                await #expect(try calculator.id == calculator.myId())
 
-            // For now, let's create a remote reference manually since resolve might return nil for remote actors
-            // We'll use the distributed actor initializer that takes an ID and system
-            let calculator = try Calculator.resolve(id: .init(), using: clientDAS)
-            await #expect(try calculator.id == calculator.myId())
+                // Now test actual XPC communication by calling methods on the resolved actor
+                // These calls should go through XPC since the actor is in a different system
+                let result1 = try await calculator.add(10, 20)
+                #expect(result1 == 30)
+                let result2 = try await calculator.multiply(6, 6)
+                #expect(result2 == 36)
 
-            // Now test actual XPC communication by calling methods on the resolved actor
-            // These calls should go through XPC since the actor is in a different system
-            let result1 = try await calculator.add(10, 20)
-            #expect(result1 == 30)
-
-            let result2 = try await calculator.multiply(5, 6)
-            #expect(result2 == 30)
+                await clientDAS.shutdown()
+            }
+            try await Task.sleep(for: .milliseconds(10))
+            await #expect(xpcHostDAS.getActiveConnectionCount() == 0)
+            #expect(xpcHostDAS.liveActorStorage.count() == 0)
         }
     }
 
@@ -92,10 +97,10 @@ struct XPCDistributedActorSystemTests {
         try await confirmation("creates remote actor just once") { confirmCreatesActor in
             let xpcHostDAS = try await XPCDistributedActorServer(
                 listener: listenerXPC,
-                actorCreationHandler: { system in
+                actorCreationHandler: { actorSystemServer in
                     // Create a Calculator actor when one isn't found for the given ID
                     confirmCreatesActor()
-                    return Calculator(actorSystem: system)
+                    return Calculator(actorSystemServer: actorSystemServer)
                 },
             )
 
@@ -130,50 +135,50 @@ struct XPCDistributedActorSystemTests {
         }
     }
 
-    // @Test("XPC AsyncStream input")
-    // func xpcAsyncStreamInputTest() async throws {
-    //     let listenerXPC = try SwiftyXPC.XPCListener(type: .anonymous, codeSigningRequirement: nil)
-    //     try await confirmation("creates remote actor just once") { confirmCreatesActor in
-    //         let xpcHostDAS = try await XPCDistributedActorServer(
-    //             listener: listenerXPC,
-    //             actorCreationHandler: { system in
-    //                 // Create a Calculator actor when one isn't found for the given ID
-    //                 confirmCreatesActor()
-    //                 return Calculator(actorSystem: system)
-    //             },
-    //         )
-
-    //         // Create a client actor system that connects to the listener's endpoint
-    //         let clientDAS = try await XPCDistributedActorClient(
-    //             connectionType: .endpoint(listenerXPC.endpoint),
-    //             codeSigningRequirement: nil,
-    //         )
-
-    //         // For now, let's create a remote reference manually since resolve might return nil for remote actors
-    //         // We'll use the distributed actor initializer that takes an ID and system
-    //         let calculator = try Calculator.resolve(id: .init(), using: clientDAS)
-    //         await #expect(try calculator.id == calculator.myId())
-
-    //         let a: Int = .random(in: 10...100)
-    //         try await confirmation("consumes input stream") {
-    //             confirmConsumeStream in
-    //             do {
-    //                 let result = try await calculator.subtractStream(
-    //                     .init { c in
-    //                         confirmConsumeStream()
-    //                         c.yield(a)
-    //                         c.finish()
-    //                     })
-    //                 #expect(result == a)
-    //             } catch {
-    //                 print("error", error)
-    //                 throw error
-    //             }
-    //         }
-    //         try await Task.sleep(for: .milliseconds(10))
-    //         await #expect(xpcHostDAS.countCodableAsyncStreams() == 0)
-    //     }
-    // }
+//     @Test("XPC AsyncStream input")
+//     func xpcAsyncStreamInputTest() async throws {
+//         let listenerXPC = try SwiftyXPC.XPCListener(type: .anonymous, codeSigningRequirement: nil)
+//         try await confirmation("creates remote actor just once") { confirmCreatesActor in
+//             let xpcHostDAS = try await XPCDistributedActorServer(
+//                 listener: listenerXPC,
+//                 actorCreationHandler: { system in
+//                     // Create a Calculator actor when one isn't found for the given ID
+//                     confirmCreatesActor()
+//                     return Calculator(actorSystem: system)
+//                 },
+//             )
+//
+//             // Create a client actor system that connects to the listener's endpoint
+//             let clientDAS = try await XPCDistributedActorClient(
+//                 connectionType: .endpoint(listenerXPC.endpoint),
+//                 codeSigningRequirement: nil,
+//             )
+//
+//             // For now, let's create a remote reference manually since resolve might return nil for remote actors
+//             // We'll use the distributed actor initializer that takes an ID and system
+//             let calculator = try Calculator.resolve(id: .init(), using: clientDAS)
+//             await #expect(try calculator.id == calculator.myId())
+//
+//             let a: Int = .random(in: 10...100)
+//             try await confirmation("consumes input stream") {
+//                 confirmConsumeStream in
+//                 do {
+//                     let result = try await calculator.subtractStream(
+//                         .init { c in
+//                             confirmConsumeStream()
+//                             c.yield(a)
+//                             c.finish()
+//                         })
+//                     #expect(result == a)
+//                 } catch {
+//                     print("error", error)
+//                     throw error
+//                 }
+//             }
+//             try await Task.sleep(for: .milliseconds(10))
+//             await #expect(xpcHostDAS.countCodableAsyncStreams() == 0)
+//         }
+//     }
 
     @Test("XPC AsyncStream early cancellation")
     func xpcAsyncStreamEarlyCancellationTest() async throws {
@@ -181,10 +186,10 @@ struct XPCDistributedActorSystemTests {
         try await confirmation("creates remote actor just once") { confirmCreatesActor in
             let xpcHostDAS = try await XPCDistributedActorServer(
                 listener: listenerXPC,
-                actorCreationHandler: { system in
+                actorCreationHandler: { actorSystemServer in
                     // Create a Calculator actor when one isn't found for the given ID
                     confirmCreatesActor()
-                    return Calculator(actorSystem: system)
+                    return Calculator(actorSystemServer: actorSystemServer)
                 },
             )
 
@@ -233,10 +238,10 @@ struct XPCDistributedActorSystemTests {
             confirmCreatesActor in
             let xpcHostDAS = try await XPCDistributedActorServer(
                 listener: listenerXPC,
-                actorCreationHandler: { system in
+                actorCreationHandler: { actorSystemServer in
                     // Create a Calculator actor when one isn't found for the given ID
                     confirmCreatesActor()
-                    return Calculator(actorSystem: system)
+                    return Calculator(actorSystemServer: actorSystemServer)
                 },
             )
 
@@ -262,6 +267,7 @@ struct XPCDistributedActorSystemTests {
 
                         let result2 = try await calculator.multiply(5, 6)
                         #expect(result2 == 30)
+                        await clientDAS.shutdown()
                         return calculator.id
                     }
                 }
@@ -273,6 +279,8 @@ struct XPCDistributedActorSystemTests {
             }
 
             #expect(ids.count == count)
+            try await Task.sleep(for: .milliseconds(10))
+            await #expect(xpcHostDAS.countCodableAsyncStreams() == 0)
         }
     }
 }
